Adam-Update_17
Added increasing Heart beat/rate depending on total amount of enemies in body!
If the total amount, an integer found on Player(gameobject) under Health-script, if it's higher than 14 and less than 25 (or 26..) then it has 2xSpeed, and if the total of enemies are higher than 24 then it has a 3xSpeed ! Feels great! :D